Criação de views. Separar os arquivos de constantes e de Views do arquivo main.
